# openapi.yaml
openapi: 3.1.0
info:
  title: ConstrSW BFF API
  description: |
    Backend for Frontend (BFF) REST API que agrega e simplifica o acesso aos microserviços do sistema ConstrSW.
    
    Esta API atua como uma camada intermediária entre o frontend e os microserviços, fornecendo:
    - Agregação de dados de múltiplos serviços
    - Simplificação de chamadas complexas
    - Redução do número de requisições necessárias
    - Padronização de respostas
    
    ## Autenticação
    Todos os endpoints (exceto `/auth/login` e `/auth/refresh`) requerem autenticação via Bearer Token (JWT).
    O token deve ser obtido através do endpoint `/auth/login` e incluído no header `Authorization: Bearer <token>`.
    
    ## Microserviços Agregados
    - OAuth (Autenticação e Autorização)
    - Employees (Funcionários)
    - Classes (Turmas)
    - Courses (Disciplinas)
    - Lessons (Aulas)
    - Professors (Professores)
    - Reservations (Reservas)
    - Resources (Recursos)
    - Rooms (Salas)
    - Students (Estudantes)
  version: 1.0.0
  contact:
    name: ConstrSW Team
    email: support@constrsw.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de desenvolvimento local
  - url: http://api:3000/api/v1
    description: Servidor interno (Docker)
...
paths:
...
  /employees:
    get:
      tags:
        - Employees
      summary: Listar funcionários
      description: Retorna uma lista de funcionários com paginação e busca
      operationId: listEmployees
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: search
          in: query
          schema:
            type: string
            description: Busca por nome, email ou documento
      responses:
        '200':
          description: Lista de funcionários
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Employees
      summary: Criar funcionário
      description: Cria um novo funcionário no sistema
      operationId: createEmployee
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreateRequest'
      responses:
        '201':
          description: Funcionário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /employees/{employeeId}:
    get:
      tags:
        - Employees
      summary: Buscar funcionário por ID
      description: Retorna os detalhes de um funcionário específico
      operationId: getEmployeeById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      responses:
        '200':
          description: Dados do funcionário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Employees
      summary: Atualizar funcionário (completo)
      operationId: updateEmployee
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdateRequest'
      responses:
        '200':
          description: Funcionário atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags:
        - Employees
      summary: Atualizar funcionário (parcial)
      operationId: patchEmployee
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdateRequest'
      responses:
        '200':
          description: Funcionário atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Employees
      summary: Deletar funcionário
      operationId: deleteEmployee
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      responses:
        '204':
          description: Funcionário deletado com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /courses:
    get:
      tags:
        - Courses
      summary: Listar disciplinas
      description: Retorna uma lista de disciplinas com filtros opcionais
      operationId: listCourses
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          schema:
            type: string
        - name: modality
          in: query
          schema:
            type: string
            enum: [PRESENCIAL, EAD, HIBRIDO]
      responses:
        '200':
          description: Lista de disciplinas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Courses
      summary: Criar disciplina
      operationId: createCourse
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreateRequest'
      responses:
        '201':
          description: Disciplina criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /courses/{courseId}:
    get:
      tags:
        - Courses
      summary: Buscar disciplina por ID
      operationId: getCourseById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '200':
          description: Dados da disciplina
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Courses
      summary: Atualizar disciplina (completo)
      operationId: updateCourse
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpdateRequest'
      responses:
        '200':
          description: Disciplina atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags:
        - Courses
      summary: Atualizar disciplina (parcial)
      operationId: patchCourse
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpdateRequest'
      responses:
        '200':
          description: Disciplina atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Courses
      summary: Deletar disciplina
      operationId: deleteCourse
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '204':
          description: Disciplina deletada com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /courses/{courseId}/materials:
    get:
      tags:
        - Courses
      summary: Listar materiais de um curso
      operationId: listCourseMaterials
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
        - name: name
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Lista de materiais
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaterialResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Courses
      summary: Adicionar material a um curso
      operationId: createCourseMaterial
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialCreateRequest'
      responses:
        '201':
          description: Material adicionado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /courses/{courseId}/materials/{materialId}:
    get:
      tags:
        - Courses
      summary: Buscar material específico
      operationId: getCourseMaterial
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
        - $ref: '#/components/parameters/MaterialId'
      responses:
        '200':
          description: Dados do material
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Courses
      summary: Atualizar material (completo)
      operationId: updateCourseMaterial
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
        - $ref: '#/components/parameters/MaterialId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialUpdateRequest'
      responses:
        '200':
          description: Material atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags:
        - Courses
      summary: Atualizar material (parcial)
      operationId: patchCourseMaterial
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
        - $ref: '#/components/parameters/MaterialId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialUpdateRequest'
      responses:
        '200':
          description: Material atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Courses
      summary: Deletar material
      operationId: deleteCourseMaterial
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
        - $ref: '#/components/parameters/MaterialId'
      responses:
        '204':
          description: Material deletado com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /courses/{courseId}/classes:
    get:
      tags:
        - Courses
      summary: Buscar turmas de um curso
      operationId: getCourseClasses
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
        - name: semester
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Lista de turmas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClassResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do endpoint /auth/login
  parameters:
...
    EmployeeId:
      name: employeeId
      in: path
      required: true
      description: ID único do funcionário
      schema:
        type: string
...
    CourseId:
      name: courseId
      in: path
      required: true
      description: ID único da disciplina
      schema:
        type: string
...
    MaterialId:
      name: materialId
      in: path
      required: true
      description: ID único do material
      schema:
        type: string
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Mensagem de erro
        status_code:
          type: integer
          description: Código HTTP do erro
...
    EmployeeResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        document:
          type: string
        position:
          type: string
        department:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EmployeeListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeResponse'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    EmployeeCreateRequest:
      type: object
      required:
        - name
        - email
        - document
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        document:
          type: string
        position:
          type: string
        department:
          type: string
    EmployeeUpdateRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        document:
          type: string
        position:
          type: string
        department:
          type: string
...
    CourseResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        modality:
          type: string
          enum: [PRESENCIAL, EAD, HIBRIDO]
        duration:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CourseCreateRequest:
      type: object
      required:
        - name
        - modality
      properties:
        name:
          type: string
        description:
          type: string
        modality:
          type: string
          enum: [PRESENCIAL, EAD, HIBRIDO]
        duration:
          type: integer
    CourseUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        modality:
          type: string
          enum: [PRESENCIAL, EAD, HIBRIDO]
        duration:
          type: integer
    ClassResponse:
      type: object
      properties:
        id:
          type: string
        classNumber:
          type: string
        year:
          type: integer
        semester:
          type: integer
        schedule:
          type: string
        course:
          type: object
          description: Disciplina associada à turma
          properties:
            id:
              type: string
              description: ID da disciplina
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    MaterialResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        url:
          type: string
          format: uri
    MaterialCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        url:
          type: string
          format: uri
    MaterialUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        url:
          type: string
          format: uri
...
  responses:
    BadRequestError:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Não autorizado - Token ausente, inválido ou expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'